pyproject.toml
[project]
name = "rss2discord"
version = "0.1.0"
description = "A simple app that moves your Rss feed to Discord."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "click>=8.1.7",
    "feedparser>=6.0.11",
    "msgpack>=1.1.0",
    "pydantic>=2.9.2",
    "pyyaml>=6.0.2",
    "requests>=2.32.3",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.9",
]
